+++ src/components/Body.js
@@ -1,32 +1,9 @@
-import React, { useEffect } from "react";
+import React from "react";
 import { RouterProvider, createBrowserRouter } from "react-router-dom";
 import Login from "./Login";
 import Browse from "./Browse";
-import { onAuthStateChanged } from "firebase/auth";
-import { auth } from "../utils/firebase";
-import { useDispatch } from "react-redux";
-import { addUser, removeUser } from "../store/userSlice";
 
 const Body = () => {
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    onAuthStateChanged(auth, (user) => {
-      if (user) {
-        const { uid, email, displayName } = user;
-        dispatch(
-          addUser({
-            uid,
-            email,
-            displayName,
-          })
-        );
-      } else {
-        dispatch(removeUser());
-      }
-    });
-  }, []);
-
   const appRouter = createBrowserRouter([
     {
       path: "/",
